中間リファクタリング その１ #17 ドメインモデルのディレクトリ名変更
