+++ packages/Models/Bot/Calculators/CalculatorInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Packages\Models\Bot\Calculators;
+
+use Packages\Models\Turn\Turn;
+
+interface CalculatorInterface
+{
+    public function __construct(Turn $turn);
+
+    public function culculate(): int;
+}
