+++ packages/Models/Player/PlayerInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Packages\Models\Player;
+
+interface PlayerInterface
+{
+    const PLAYER_TYPE_PREFIX_PERSON = 'P';
+    const PLAYER_TYPE_PREFIX_BOT    = 'B';
+
+    public function getid(): string;
+    public function getName(): string;
+    public function getPlayerType(): string;
+
+    public function isPerson(): bool;
+    public function isBot(): bool;
+}
