+++ tests/Feature/Color/ColorTest.php
@@ -4,7 +4,7 @@ namespace Tests\Feature\Color;
 
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Foundation\Testing\WithFaker;
-use Packages\Domain\Color\Color;
+use Packages\Models\Board\Color;
 use Tests\TestCase;
 
 class ColorTest extends TestCase
@@ -16,8 +16,8 @@ class ColorTest extends TestCase
      */
     public function testOpposite()
     {
-        $black = new Color(Color::COLOR_BLACK);
-        $white = new Color(Color::COLOR_WHITE);
+        $black = new Color(Color::COLOR_CODE_BLACK);
+        $white = new Color(Color::COLOR_CODE_WHITE);
 
         $this->assertSame($black->opposite()->toCode(), $white->toCode());
         $this->assertSame($white->opposite()->toCode(), $black->toCode());
